moved utils to folder
