+++ services/dask-sidecar/src/simcore_service_dask_sidecar/utils/dask.py
@@ -19,7 +19,7 @@ from models_library.progress_bar import ProgressReport
 from models_library.rabbitmq_messages import LoggerRabbitMessage
 from servicelib.logging_utils import LogLevelInt, LogMessageStr, log_catch
 
-from .rabbitmq_plugin import get_rabbitmq_client
+from ..rabbitmq_plugin import get_rabbitmq_client
 
 _logger = logging.getLogger(__name__)
