+++ services/dask-sidecar/src/simcore_service_dask_sidecar/worker.py
@@ -16,9 +16,13 @@ from settings_library.s3 import S3Settings
 
 from ._meta import print_dask_sidecar_banner
 from .computational_sidecar.core import ComputationalSidecar
-from .dask_utils import TaskPublisher, get_current_task_resources, monitor_task_abortion
 from .rabbitmq_plugin import RabbitMQPlugin
 from .settings import ApplicationSettings
+from .utils.dask import (
+    TaskPublisher,
+    get_current_task_resources,
+    monitor_task_abortion,
+)
 from .utils.logs import setup_app_logging
 
 _logger = logging.getLogger(__name__)
